pyproject.toml
[project]
name = "cartographer"
version = "0.1.0"
description = ""
authors = [
    {name = "avishek-sen-gupta",email = "asgupta@thoughtworks.com"}
]
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "langgraph (>=1.0.7,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "neo4j (>=6.1.0,<7.0.0)"
]

[project.optional-dependencies]
ml = [
    "llama-cpp-python (>=0.3.0)",
    "huggingface-hub (>=0.25.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
mypy = "^1.19.1"
pytest = "^9.0.2"
coverage = "^7.13.2"
neo4j = "6.1.0"
python-dotenv = "1.2.1"
tree-sitter-language-pack = "0.13.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
